+++ src/ska_low_mccs_daq/daq_receiver/daq_receiver_device.py
@@ -284,9 +284,11 @@ class MccsDaqReceiver(MccsBaseDevice):
         self._rms_plot: np.ndarray
         self._nof_saturations: int
         self._nof_packets: int
+        self._nof_samples: int
         self._data_rate: float
         self._receive_rate: float
         self._drop_rate: float
+        self._correlator_time_taken: float
         self._skuid_url: str
 
     def init_device(self: MccsDaqReceiver) -> None:
@@ -353,9 +355,11 @@ class MccsDaqReceiver(MccsBaseDevice):
         self._rms_plot = np.zeros(shape=(256, 512), dtype=float)
         self._nof_saturations = 0
         self._nof_packets = 0
+        self._nof_samples = 0
         self._receive_rate = 0
         self._drop_rate = 0
         self._data_rate = 0
+        self._correlator_time_taken = 0.0
         self.set_change_event("healthState", True, False)
         self.set_archive_event("healthState", True, False)
 
@@ -457,12 +461,16 @@ class MccsDaqReceiver(MccsBaseDevice):
             self._device.set_archive_event("nofSaturations", True, False)
             self._device.set_change_event("nofPackets", True, False)
             self._device.set_archive_event("nofPackets", True, False)
+            self._device.set_change_event("nofSamples", True, False)
+            self._device.set_archive_event("nofSamples", True, False)
             self._device.set_change_event("dataRate", True, False)
             self._device.set_archive_event("dataRate", True, False)
             self._device.set_change_event("receiveRate", True, False)
             self._device.set_archive_event("receiveRate", True, False)
             self._device.set_change_event("dropRate", True, False)
             self._device.set_archive_event("dropRate", True, False)
+            self._device.set_change_event("correlatorTimeTaken", True, False)
+            self._device.set_archive_event("correlatorTimeTaken", True, False)
 
             return (ResultCode.OK, "Init command completed OK")
 
@@ -572,6 +580,14 @@ class MccsDaqReceiver(MccsBaseDevice):
         self._nof_packets = 0
         self.push_change_event("nofPackets", 0)
         self.push_archive_event("nofPackets", 0)
+
+        self._nof_samples = 0
+        self.push_change_event("nofSamples", 0)
+        self.push_archive_event("nofSamples", 0)
+
+        self._correlator_time_taken = 0.0
+        self.push_change_event("correlatorTimeTaken", 0)
+        self.push_archive_event("correlatorTimeTaken", 0)
 
     def _received_data_callback(
         self: MccsDaqReceiver,
@@ -1166,20 +1182,38 @@ class MccsDaqReceiver(MccsBaseDevice):
     @attribute(dtype="DevLong")
     def nofSaturations(self: MccsDaqReceiver) -> int:
         """
-        Return the nof saturations of the last station beam consumer integration.
+        Return the nof saturations of the last consumer integration.
 
-        :return: the nof saturations of the last station beam consumer integration.
+        :return: the nof saturations of the last consumer integration.
         """
         return int(self._nof_saturations)
 
     @attribute(dtype="DevLong")
     def nofPackets(self: MccsDaqReceiver) -> int:
         """
-        Return the nof packets of the last station beam consumer integration.
+        Return the nof packets of the last consumer integration.
 
-        :return: the nof packets of the last station beam consumer integration.
+        :return: the nof packets of the last consumer integration.
         """
         return int(self._nof_packets)
+
+    @attribute(dtype="DevLong")
+    def nofSamples(self: MccsDaqReceiver) -> int:
+        """
+        Return the nof samples of the last consumer integration.
+
+        :return: the nof packets of the last consumer integration.
+        """
+        return int(self._nof_samples)
+
+    @attribute(dtype="DevFloat")
+    def correlatorTimeTaken(self: MccsDaqReceiver) -> float:
+        """
+        Return the time taken for the last correlation.
+
+        :return: the time taken for the last correlation.
+        """
+        return self._correlator_time_taken
 
     @attribute(dtype="DevFloat")
     def dataRate(self: MccsDaqReceiver) -> float | None: